wip: shake rules for opam
Setting up rocq is turning out to be quite the headache due to
switches. These are opam's take on virtual environments, and
work almost entirely based off of environment variables, which
are not particularly pleasant to deal with in a build script!

To deal with this, I've factored out a lot of opam specific stuff
into a bunch of shake rules: hopefully this will make adding future
proof assistants less horrible.